mousedev: mousedev patch for qemumips
1/1 [
Author: Richard Griffiths
Email: richard.griffiths@windriver.com
Subject: mousedev: mousedev patch for qemumips
Date: Tue, 21 Sep 2010 13:22:55 -0700

The consumers of the mousedev data in MIPS uperspace assume
that the data is little endian. So if we are running on a
big endian CPU we force the output data to be little endian.

Note: this may be reverted if/when the consuming devices
add options, or options are found that can byte swap the
data in userspace. This is simply a commit of convienience
to demonstrate an issue.

Signed-off-by: Richard Griffiths <richard.griffiths@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>