ALSA: seq/oss - Fix double-free at error path of snd_seq_oss_open()
CQID: 236851 236856 236857

1/3 [
Author: Takashi Iwai
Email: tiwai@suse.de
Subject: ALSA: seq/oss - Fix double-free at error path of snd_seq_oss_open()
Date: Sat, 16 Oct 2010 22:44:01 +0800

commit 27f7ad53829f79e799a253285318bff79ece15bd from upstream

The error handling in snd_seq_oss_open() has several bad codes that
do dereferecing released pointers and double-free of kmalloc'ed data.
The object dp is release in free_devinfo() that is called via
private_free callback.  The rest shouldn't touch this object any more.

The patch changes delete_port() to call kfree() in any case, and gets
rid of unnecessary calls of destructors in snd_seq_oss_open().

Fixes CVE-2010-3080.

Reported-and-tested-by: Tavis Ormandy <taviso@cmpxchg8b.com>
Cc: <stable@kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Integrated-by:  Shenbo Huang<shenbo.huang@windriver.com>
]

2/3 [
Author: Steven Rostedt
Email: srostedt@redhat.com
Subject: tracing: Do not allow llseek to set_ftrace_filter
Date: Sat, 16 Oct 2010 22:05:17 +0800

Commit 9c55cb12c1c172e2d51e85fbb5a4796ca86b77e7 from upstream.

Reading the file set_ftrace_filter does three things.

1) shows whether or not filters are set for the function tracer
2) shows what functions are set for the function tracer
3) shows what triggers are set on any functions

3 is independent from 1 and 2.

The way this file currently works is that it is a state machine,
and as you read it, it may change state. But this assumption breaks
when you use lseek() on the file. The state machine gets out of sync
and the t_show() may use the wrong pointer and cause a kernel oops.

Luckily, this will only kill the app that does the lseek, but the app
dies while holding a mutex. This prevents anyone else from using the
set_ftrace_filter file (or any other function tracing file for that matter).

A real fix for this is to rewrite the code, but that is too much for
a -rc release or stable. This patch simply disables llseek on the
set_ftrace_filter() file for now, and we can do the proper fix for the
next major release.

Reported-by: Robert Swiecki <swiecki@google.com>
Cc: Chris Wright <chrisw@sous-sol.org>
Cc: Tavis Ormandy <taviso@google.com>
Cc: Eugene Teo <eugene@redhat.com>
Cc: vendor-sec@lst.de
Cc: <stable@kernel.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Integrated-by:  Shenbo Huang<shenbo.huang@windriver.com>
]

3/3 [
Author: Dan Rosenberg
Email: dan.j.rosenberg@gmail.com
Subject: xfs: prevent reading uninitialized stack memory
Date: Sat, 16 Oct 2010 21:37:16 +0800

Commit a122eb2fdfd78b58c6dd992d6f4b1aaef667eef9 from upstream.

The XFS_IOC_FSGETXATTR ioctl allows unprivileged users to read 12
bytes of uninitialized stack memory, because the fsxattr struct
declared on the stack in xfs_ioc_fsgetxattr() does not alter (or zero)
the 12-byte fsx_pad member before copying it back to the user.  This
patch takes care of it.

Signed-off-by: Dan Rosenberg <dan.j.rosenberg@gmail.com>
Reviewed-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Alex Elder <aelder@sgi.com>
Integrated-by:  Shenbo Huang<shenbo.huang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>