sctp: Fix out-of-bounds reading in sctp_asoc_get_hmac()
CQID: 237931

1/1 [
Author: Dan Rosenberg
Email: drosenberg@vsecurity.com
Subject: sctp: Fix out-of-bounds reading in sctp_asoc_get_hmac()
Date: Sun, 24 Oct 2010 14:04:18 +0800

Commit 51e97a12bef19b7e43199fc153cf9bd5f2140362 from upstream.

The sctp_asoc_get_hmac() function iterates through a peer's hmac_ids
array and attempts to ensure that only a supported hmac entry is
returned.  The current code fails to do this properly - if the last id
in the array is out of range (greater than SCTP_AUTH_HMAC_ID_MAX), the
id integer remains set after exiting the loop, and the address of an
out-of-bounds entry will be returned and subsequently used in the parent
function, causing potentially ugly memory corruption.  This patch resets
the id integer to 0 on encountering an invalid id so that NULL will be
returned after finishing the loop if no valid ids are found.

Signed-off-by: Dan Rosenberg <drosenberg@vsecurity.com>
Acked-by: Vlad Yasevich <vladislav.yasevich@hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Integrated-by:  Shenbo Huang<shenbo.huang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>