filesystems: ensure we call out EXT3, EXT4 and bart fs.
We've largely not updated the baseline config for a while,
since we got EXT3 for free by the default=y in the Kconfig.
Call it out explicitly, and while we are there, lets get
with the times and have EXT4 and BTRFS enabled as well.
The addition of JBD2 is purely a result of dependencies.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>