Cavium Octeon patches for 3.0.4 from sustaining pull 10.11.19
WIND00237003

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: Fix boot failed on cavium target when CVMX_ENABLE_DFA_FUNCTIONS is defined
Date: Fri, 22 Oct 2010 18:27:35 +0800

This issue is caused by the inconsistent define of CVMX_FPA_OUTPUT_BUFFER_POOL
on driver code and on simple_exec_open code when CVMX_ENABLE_DFA_FUNCTIONS is defined.
on driver code, CVMX_FPA_OUTPUT_BUFFER_POOL is defined as 2, however, when
CVMX_ENABLE_DFA_FUNCTIONS is defined, CVMX_FPA_OUTPUT_BUFFER_POOL is defined as 3 on simple_exec_open
in generated cvmx-config.h file. The point is that FPA pool is initialized on driver code,
so the driver code has not initialized enough buffer for PKO queue command buffers.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

WIND00227970

1/1 [
Author: Yong Zhang
Email: yong.zhang@windriver.com
Subject: octeon-ethernet: Fix race freeing transmit buffers
Date: Tue, 24 Aug 2010 14:36:11 +0800

commit a620c1632629b42369e78448acc7b384fe1faf48 upstream

The existing code had the following race:

Thread-1                       Thread-2

inc/read in_use
                               inc/read in_use
inc tx_free_list[qos].len
                               inc tx_free_list[qos].len

The actual in_use value was incremented twice, but thread-1 is going
to free memory based on its stale value, and will free one too many
times.  The result is that memory is freed back to the kernel while
its packet is still in the transmit buffer.  If the memory is
overwritten before it is transmitted, the hardware will put a valid
checksum on it and send it out (just like it does with good packets).
If by chance the TCP flags are clobbered but not the addresses or
ports, the result can be a broken TCP stream.

The fix is to track the number of freed packets in a single location
(a Fetch-and-Add Unit register).  That way it can never get out of sync
with itself.

We try to free up to MAX_SKB_TO_FREE (currently 10) buffers at a time.
If fewer are available we adjust the free count with the difference.
The action of claiming buffers to free is atomic so two threads cannot
claim the same buffers.

Signed-off-by: David Daney <ddaney@caviumnetworks.com>
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
[Yong Zhang: fix patch confict and build]
Integrated-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>