powerpc: AMP: make the parameter of dtc's -b option changeable
1/1 [
Author: Wu Zhangjin
Email: zhangjin.wu@windriver.com
Subject: powerpc: AMP: make the parameter of dtc's -b option changeable
Date: Wed, 28 Jul 2010 16:20:12 +0800

Before, the parameter of dtc's -b option is hardcoded as 0 in the Makefile, but
in AMP or some other situations, this may need to be changeable.  for example,
to make the core1 of fsl_p1020 run well with the dts file, we must set the boot
cpu as 1 via compiling the dts file with "-b 1" of dtc.

This patch makes the parameter of the -b option of dtc changeable via adding a
DTB_BOOT_CPU for the Makefile, if no explicit change pass, it is reserved as 0.

Usage:

$ cd /patch/to/linux-build-directory/
$ make p1020rdb_camp_core1.dtb DTB_BOOT_CPU=1

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>