add intel_core_qm67_pch async. 4.0 BSP
support -standard and -preempt-rt ktype, this is an update of 4.0 on
tree BSP, but use a different BSP name, the old name is
intel_core_q67_pch while the new name is intel_core_qm67_pch, so add a
suite of new teplates for new name, and need too remove the old name
template suite.

This update includes one patch, i pushed it down to standard.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>