p4080/gpio: export all the GPIO registers in DTS
1/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: p4080/gpio: export all the GPIO registers in DTS
Date: Tue, 14 Dec 2010 13:51:00 +0800

Currently only part of the GPIO registers are exported via DTS.
And the registers GPDIR and GPODR are missed. But in some cases
we need to access them. So export all the GPIO registers in DTS.
Also adjust the p4080ds mdio driver for this change.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: p4080ds/mdio: set direction of the GPIO ports used by mdio multiplex
Date: Tue, 14 Dec 2010 14:50:52 +0800

The GPIOs are used to control the MDIO bus multiplexers for both
SGMII and XAUI. In general the bootloader will set the direction
of the GPIO ports as output. We only need to set the correct value
to the GPDAT register when accessing the MDIO bus. But this is not
true for the bootrom. It only set the direction type for SGMII,
and leave the direction of the ports for XAUI uninitialized.
Here we just set the correct direction type for both SGMII and XAUI
no matter whether the bootloader already set it or not. It is no harm,
and would avoid the ugly ifdef.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

3/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080: introduce a function to check the SOC version
Date: Tue, 14 Dec 2010 15:49:05 +0800

In some cases we need to check the SOC version. So introduce
a function for that.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

4/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080ds: workaround for errata gen8
Date: Tue, 14 Dec 2010 16:57:54 +0800

Due to the errata gen8, the XAUI MDIO bus is powered down when
SERDES protocl 0x10 is used. We have to use fixed link phy in
this situation. This errata only affect rev1, and is fixed in
rev2.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

5/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080ds: workaround for errata serdes9
Date: Tue, 14 Dec 2010 10:10:00 +0800

Workaround for errata serdes9. This is based on the fix in FSL
p4080 SDK 2.2 u-boot.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>