kernel: use SLAB allocator with NUMA
CQID: 243493

1/1 [
Author: Greg Moffatt
Email: greg.moffatt@windriver.com
Subject: kernel: use SLAB allocator with NUMA
Date: Tue, 21 Dec 2010 09:15:02 -0500

When NUMA is configured for a CGL kernel on an intel 5500 target,
the kernel panics when init is started.  There is a conflict between
NUMA, GRSEC/PAX, and the SLUB allocator.

To address this the SLAB allocator should be configured instead.  To
achieve this the Kconfig for the allocator is changed such that if
NUMA is configured by a user the SLAB allocator is selected, rather
than the SLUB allocator.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>