WRHV/e500: ASID kconfig
1/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/e500: ASID kconfig
Date: Mon, 31 Jan 2011 13:43:34 -0500

ASID - Address Space Identifier

The main purpose of ASID is to act as a MMU Assist register
thereby limiting the number of flushes and increasing the
overall guest performance.

Creation of user selectable and tunable ASID value. This
menu item enables and disables the ASID performance
feature.

Ideally we could have enabled and set the number of ASID's
through the bootline.  But the call to parse the bootline
value comes way too late for this.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

2/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV: Increase exception vector size
Date: Mon, 14 Feb 2011 15:10:41 -0500

When SMP is enabled on an E500 target the exception
vector was not large enough.   Increase the size so
that when extra instructions are added the whole
intended exception is present as intended.

[Context mods to deal with direct IRQ]
Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

3/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV: Export status page
Date: Mon, 14 Feb 2011 15:39:28 -0500

Previously the status page was not being exported
which did not pose a problem.  But now the status
page is needed during:
context creation, destruction, switching and MMU
restoration.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

4/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV: Remove unneeded and obsolete code
Date: Tue, 15 Feb 2011 01:15:45 -0500

No longer used and obsoleted in the HV many moons ago.

This commit should be merged into this:

--------------------
commit 3916ed22ec5c0d31ab6ced4dbb86cd5326934c91
Author: WRS Support <support@windriver.com>
Date:   Tue Apr 27 15:22:48 2010 +0800

    powerpc wrhv: introduce powerpc specifics of WR guest

    These powerpc specific additons are the Linux specific
    additions that were not a part of the reference VBI
    implemenation.

    Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
    Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>

--------------------

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

5/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/ASID: paravirt context switch
Date: Fri, 18 Feb 2011 14:51:55 -0500

Paravirtualization of context creation, destruction, switching
is needed such that within each the correct VBI calls are made.
As well the total number of contexts/ASID's must be limited to
the maximum available ASID's.

The kernel needs to limit the total number of ASID's available,
this cap is controlled through the use of a kernel config
option.  (CONFIG_WRHV_NUM_ASID)

When the kernel first boots it creates an ASID of 1.  Upon
successfully booting user processes need to start at context
ID of 2 to MAX_ASID-1.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

6/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/ASID: Associate VMMU Handle and ASID to every MM
Date: Tue, 15 Feb 2011 21:34:09 -0500

There needs to be a direct corelation between every
MM and VMMU Handle.  A VMMUHandle exists for the duration
of every allocated MM.  This association is used during
context creation and deletion.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

7/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/ASID: proper config and control setup before sys_ctx_load
Date: Wed, 16 Feb 2011 00:25:33 -0500

The vbi_load_ctx stub that is included as part of the official
Hypervisor DVD is insufficient.  Prior to calling sys_ctx_load,
the config and control page need to be updated with the proper
values so that when the HV calls RFI Linux returns to the correct
location.

Currently the only caller of this function is the ASID performance
optimization.

This code is owned by the VBI and shortly will become part of the
official VBI release at which point this patch can be reverted.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

8/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/ASID: Update config and control struct in and out of exceptions
Date: Fri, 18 Feb 2011 14:59:36 -0500

Linux needs to push the virtual ASID and VMMU Handle when
returning from and going to exceptions so that they are
in sync with the Hypervisor.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

9/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: WRHV/ASID: Paravirtualized set_context adjustments
Date: Wed, 16 Feb 2011 10:52:22 -0500

The paravirtualized version of set_context is insuffifcient
for controlling the Hypervisor with ASID optimizations enabled.
Function signature updates are needed to distinguish between
non-asid enabled hypervisor and enabled, this provides backwards
compatibility.

The first ASID is treated differently within the hypervisor,
given that it is statically mapped within the hypervisor.

Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
]

10/10 [
Author: Jeremy McNicoll
Email: jeremy.mcnicoll@windriver.com
Subject: Update the vbi to Nov 19
Date: Fri, 18 Feb 2011 15:24:07 -0500

This is a preliminary patch for testing.

Minor MMU config and control changes from Feb 11.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>