Cavium: Hotplug CPU modifications
1/2 [
Author: ltian
Email: le.tian@windriver.com
Subject: Cavium: Hotplug CPU modifications
Date: Fri, 12 Nov 2010 15:46:45 +0800

Source: SDK 2.0.0-366

Cleanup hotplug cpu code:

For CONFIG_HOTPLUG_CPU, we set the 'possible' cpus to the set of all
cores present, even if they are not in the boot_coremask.

Update the avail_coremask when the core is plugged in, so that it will
not be available for spawning other applications.

Improve HOTPLUG_CPU related panic messages

Signed-off-by: ltian <le.tian@windriver.com>
]

2/2 [
Author: ltian
Email: le.tian@windriver.com
Subject: Finish randomize mmap for non-cgl configuration
Date: Mon, 15 Nov 2010 14:56:57 +0800

Source: Cavium SDK 2.0.0-366, Split from previous patch

When we apply BSP patches, some of the kernel source files we start
with are different between CGL and standard kernels. In most cases, we
either don't care about the differences, or they don't affect the
files we get in a way that the BSP patches can't be applied.

The case of arch/mips/kernel/syscall.c is slightly different,
however. There is a conflict in that a one-line change that the BSP
patches want to apply has already been applied in the CGL kernel. This
means that the BSP patch fails.

This separates the offending patch hunk so that it can be
conditionally included, depending on whether this is a CGL build or
not.

Signed-off-by: ltian <le.tian@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>