wrhv/p4080: embed the FMAN firmware into the kernel
CQID: WIND00257968

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080: embed the FMAN firmware into the kernel
Date: Tue, 1 Mar 2011 18:29:59 +0800

We need load the firmware first before using the FMAN. But at present
flash is not supported on the p4080 guest OS. So embed the firmware
into the kernel.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080: use the general GPIO function to set the multiplex of MDIO
Date: Wed, 2 Mar 2011 10:05:10 +0800

Previously we set the GPIO directly in our MDIO driver. But now we use
the general GPIO lib function to do this work in the new MDIO driver.
So adjust the DTS according to this change.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>