wrhv/ppc: fix a bug in mapping vb configuration pages
CQID: WIND00258378

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/ppc: fix a bug in mapping vb configuration pages
Date: Fri, 4 Mar 2011 02:49:33 +0800

According to the comments previously we don't know the size of the
object we mapped. Then we just assume that it is just use one page.
But things have changed. Now we already got the object size,
and in some cases it is beyond one page size. So we should take
into account the size when mapping.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/virtio: don't run input vq cb before we fill the read buffer
Date: Fri, 4 Mar 2011 10:12:00 +0800

It makes no sense to run the input virtual queue callback before
we fill the read buffer. It is also possible that some members of
ports_device struct are still not initialized. So running the callback
will cause panic in this situation.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>