RT: bugfixes and cosmetic changes from upstream alignment
Now that 34-rt exists externally we are aligning the work Mark
did earlier to that, by examining the deltas and seeing whether
they are justififed, and what their impact would be.

This commit represents the cumulative work of two RRs that
audited and align the majority of our old code with that of
the upstream.  Once we are completely aligned, we'll make the
final cosmetic step of swapping in the upstream directly so
that we have the complete commit history in addition to code
alignment.

That of course means that the commit logs in the associated
patches added here in this commit are only transitory.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>