preempt-irq: prevent softirq thread from dropping its binding
1/1 [
Author: Mark Asselstine
Email: mark.asselstine@windriver.com
Subject: preempt-irq: prevent softirq thread from dropping its binding
Date: Tue, 6 Jul 2010 13:13:06 -0700

When softirq threads are created they are each bound to a
specific CPU on SMP systems. Thread creation takes place before
CPU may become online and it is also possible for softirqs to
be raised prior to a CPU becoming online. If we allow
a thread to be woken before the CPU it is bound to is online it
will drop this binding and bind to CPU0. This will lead to a mismatch
situation with softirqdata and cause all sort of issues at boot.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>