wrhv/ppc: don't migrate the NOBALANCING irq when a cpu down
1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/ppc: don't migrate the NOBALANCING irq when a cpu down
Date: Wed, 13 Apr 2011 09:52:48 +0800

When a cpu is down we will migrate the irq on it to other cpu.
But if an irq is explicitly set with IRQF_NOBALANCING flag,
it mean that we don't want this irq to be handled on other cpu.
So just leave it as is.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>