wrhv/p4080: enable hard FPU on rev2 board
CQID: WIND00270347

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/ppc: fix the limitation of using fixed address for each IVOR
Date: Mon, 25 Apr 2011 15:48:28 +0800

Currently each interrupt handler use a fixed length space, and we
will set the GOS IVOR based on this assumption. This has the following
drawbacks:

1. Create bigger footprint. Currently the size of text section is
7743088, after using unfixed length handler the size of text section
is reduced to 7736432.

2. It is very inconvenient to adjust the interrupt handler according
to the cpu type.

So instead of assuming every interrupt handler is at the fixed address,
we will set the address of each interrupt handler to exception table
before setting the GOS IVOR via vbi.

It is also no harm to set the direct interrupt handler even that is not
enabled. So always set it.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/p4080: enable hard FPU on rev2 board
Date: Mon, 25 Apr 2011 17:19:27 +0800

Due to the errata CPU12 the core may stall when executing a
floating-point instruction on rev1 board, but it has been fixed
in rev2 board. So add a dynamic check to use FPU on rev2 board,
and still do software emulation for floating-point instruction
on rev1 board.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>