wrhv/e500mc: fix several bugs in __tlbil_va
CQID: WIND00272232

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/e500mc: fix several bugs in __tlbil_va
Date: Fri, 6 May 2011 08:52:01 +0800

There are several bugs in __tlbil_va when MMU_FTR_USE_TLBILX feature
is not set:

* The opcode of "tlbsx 0,r4" should be 0x7c002724.

* We don't need to set the MAS3 before executing tlbwe instruction
because the hypervisor already put the guest physical address to
MAS3 after tlbsx. This also fixes a undefined reference error to
"raddr" if PCI is disabled.

* Currently it will skip over the invalidation of the TLB entry in
AS 0 if we don't find a valid TLB entry in AS 1. This is not correct.
We should definitely invalidate the TLB entries in both AS 0 and 1.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

This fix above is squashed with the original code:
  WRHV-fsl_p4080-Implement-tlb-invalidate-function.patch

Drop the following patch as well because the code is made obsolete by
the fix above:
  WRHV-fsl_p4080-Export-the-real-start-address-of-VB.patch

The context of this fix above contained the fix of the following patch,
so squash this one too while we are here:
   ppc-mc-Fix-incorrect-CR0-setting.patch

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>