rt: mmc deadlock fix
CQID: WIND00273243

 ----------
	weiwei.wang@windriver.com

This is a dead lock issue introduced by:
commit 569bed2ab622cde48bdb0ebae3da24426751b024
Author: Andrew Liu <shengping.liu@windriver.com>
Date:   Tue Aug 3 20:14:52 2010 +0800

    rt_optimize: MMC replace local_irq by spin lock call

    Signed-off-by: Herve Patriarche <herve.patriarche@windriver.com>
    Integrated-by: Andrew Liu <shengping.liu@windriver.com>

The intention of this patch is to use spin_lock_irq* to instead of
local_irq_save, and it will reduce the times of irq disabled because
spin_lock_irq* will not really disable local irq. But this way introduces
dead lock:
sdhci_irq
  -> spin_lock(&host->lock);
  -> ...
  -> sdhci_read_block_pio -> spin_lock_irqsave(&host->lock, flags);

deadlock appears.

Actual, to implement the optimization, local_irq_*_nort is the right one
to do so, and it will not disable local irq in preempt_rt kernel neither.

Here 2 patches attached:
0001-Revert-rt_optimize-MMC-replace-local_irq-by-spin-loc.patch
This is the revert patch. we can just delete the reverted patch when rebuild
our git tree.
 ----------

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>