wrhv/ppc: introduce memmap kernel parameter
CQID: WIND00273230

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: wrhv/ppc: introduce memmap kernel parameter
Date: Tue, 10 May 2011 14:44:23 +0800

On PPC GOS some memories are assigned to some devices for private use
by bootloader, such as the memory exclusively used by QMAN and BMAN
on p4080 board. So we need a method to reserve these memories.
We can't do this in the corresponding drivers because it is too
late for this purpose. The kernel parameter "memmap=nn[KMG]$ss[KMG"
is already used by x86 architecture for this function. So reuse it
on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>