WRHV: Cavium: Accommodate Hypervisor archTickDeliver changes
CQID: WIND00273626

1/1 [
Author: Phil Staub
Email: Phil.Staub@windriver.com
Subject: WRHV: Cavium: Accommodate Hypervisor tick interrupt changes
Date: Wed, 11 May 2011 08:57:57 -0700

Hypervisor was modified to route the tick interrupt through the
vioapic. This means we no longer directly poll the summary register
for the tick interrupt.

Signed-off-by: Phil Staub <Phil.Staub@windriver.com>
]

This fix above is squashed with commit:
  WRHV: Cavium: Paravirtualize IRQ dispatch handler

Squash the following with it as well, the context is shared:
  WRHV: Cavium: Rework IRQ dispatch to handle the VIOAPIC interrupts

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>