bman: wake up the process waiting for the bman event
CQID: WIND00275259

This is a revisit of an old patch, WIND00232148.

This is a defect that was fixed by Kexin, but the fix was then
overridden by SDK 2.2 from FSl vendor.

Move commit "bman: wake up the process waiting for the bman event" so
that it follows "BMAN: Update Bman subsystem".

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>