fsl_p4080-mc: replace with TLB1 entry to map UART REG space
CQID: WIND00276611

1/1 [
Author: Tiejun Chen
Email: tiejun.chen@windriver.com
Subject: fsl_p4080-mc: replace with TLB1 entry to map UART REG space
Date: Mon, 16 May 2011 15:11:59 +0800

The hypervisor only saves the persistent TLB1 entries but NOT TLB0,
the TLB0 entries are by nature a cache meant to refreshed when needed
via an MMU TLB miss handling. For example we don't save TLB0 entries
when you switch between processes in Linux. For P4080 LPID tag separates
the TLB's between VBs in a similar way PID does for processes but
they use the same cache.

So a TLB0 entry may be replaced by a more recent access to a TLB
which has similar address attributes as same 'way'. This is done by
the hardware and we have no control over it.

We have to use TLB1 entry to replace the original TLB0 entry to map
UART REG space for early output. And the hypervisor always guarantee TLB1
to make sure that VB don't corrupt any valid TLB1 entry. And the hypervisor
can search the free TLB1 entry index to replace the index we set. So we can
set the TLB1 index with any value and here we use '0' by default. And while
bootstrap TLB1 entries are enough for VB so its also safe.

And use PPC_TLBILX_ALL to invalidate all unnecessary TLB entries in current
partititon.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>