fsl_p1022_p1013: disable JOG mode for WB debug via JTAG
CQID: WIND00277876

1/1 [
Author: Tiejun Chen
Email: tiejun.chen@windriver.com
Subject: fsl_p1022_p1013: disable JOG mode for WB debug via JTAG
Date: Mon, 23 May 2011 10:20:44 +0800

The JTAG debug has limitations comparing running the OS with full
functionalities for the core. The "WR_OCD_DEBUG" flag is for
adding/removing certain features to enable background debug being
functional. We especially have difficulty to fully control the CPU
when it's in power saving mode. For some CPU family we have to take
the whole power management module out of the system to enable JATG
debugging. For this specific case, when the CPU in jogging mode,
the JTAG detect it as "STOPPED" mode then we put in "BKM" mode
instead of "RUN" mode. And there is no way for the JTAG to
tell the CPU in JOG mode. So we have to disable JOG only when we
use WorkBench to debug this target via JTAG.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>