Fix merge regression from external kdb to upstream kdb
1/1 [
Author: Jason Wessel
Email: jason.wessel@windriver.com
Subject: Fix merge regression from external kdb to upstream kdb
Date: Wed, 21 Jul 2010 19:27:06 -0500

In the process of merging kdb to the mainline, the kdb lsmod command
stopped printing the base load address of kernel modules.  This is
needed for using kdb in conjunction with external tools such as gdb.

Simply restore the functionality by adding a kdb_printf for the base
load address of the kernel modules.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>