powerpc: Add Kconfig support for PPC_MERGE
1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: powerpc: Add Kconfig support for PPC_MERGE
Date: Thu, 26 May 2011 10:32:19 +0800

commit dbc1c5c250cbedccf3571597d156e581e34b2944 remove Kconfig
support for PPC_MERGE, it states:
"There are no users of PPC_MERGE in tree so we can get rid of it."
This is not true because ocf/talitos still using it, so for ocf/talitos
working, we should add this kernel config.
Another method for ocf/talitos working is modify talitos source code,
but clearly it's not proper.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>