net/gianfar: use eth0 as the default interface when ifname is not set
CQID: WIND00281966

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: net/gianfar: use eth0 as the default interface when ifname is not set
Date: Mon, 13 Jun 2011 13:45:40 +0800

Use the eth0 as the default interface when ifname is not set.
Otherwise we will overwrite the stack frame when copying the
content of wrhv_net_name to a local variable because we use
"strlen(wrhv_net_name) - 1" as the length of the string.
Also add a check to validate the mac address we set in
command line.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>