udp/recvmsg: Clear MSG_TRUNC flag when starting over for a new packet
CQID: WIND00283708 WIND00282432

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: udp/recvmsg: Clear MSG_TRUNC flag when starting over for a new packet
Date: Mon, 20 Jun 2011 10:15:18 +0800

Consider this scenario: When the size of the first received udp packet
is bigger than the receive buffer, MSG_TRUNC bit is set in msg->msg_flags.
However, if checksum error happens and this is a blocking socket, it will
goto try_again loop to receive the next packet, but if the size of the
next udp packet is smaller than receive buffer, MSG_TRUNC flag should not
be set, but because MSG_TRUNC bit is not cleared in msg->msg_flags before
receive the next packet, MSG_TRUNC is still set, this is wrong.

Fix this problem by clearing MSG_TRUNC flag when starting over for a
new packet.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>