wrhv: fixup some stuff found building randconfig
A quick overview of the changes:

* move DEFINE_PER_CPU out of functions, this prevents warnings
like "error: section attribute cannot be specified for local variables"

* add missing asm/fixmap.h, preventing a build error for missing
set_fixmap() definition, must be getting it indirectly using other
config settings.

* remove check for CONFIG_PPC85xx_VT_MODE around code which
is known to be redefined in arch/powerpc/include/asm/pv_pgtable-ppc32.h
this matches the check in that header file.

* fixed an obvious mistaken struct name in a CONFIG_DEBUG_VIRTUAL_IRQS
block.

* make CONFIG_WRHV depend on not being X86 and SPARSE_IRQ, this
simply will not build as is. If we want SPARSE_IRQ support on
X86 we need to do additional work.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>