For standard, features/ocf - sustaining pull 11-07-21
CQID: WIND00265642

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: ocf-talitos: modify talitos_match so as to keep consistent with dts
Date: Fri, 13 May 2011 12:11:06 +0800

The match_table for ocf talitos driver is not consistent with dts,
this makes talitos_probe() never be called, so talitos driver
cannot work.

Modify entries of talitos_match so as to make it keep consistent with dts.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>