async bsp merge 110802: patches for standard
CQID: WIND00280927

1/5 [
Author: Alexander Graf
Email: agraf@suse.de
Subject: PPC: Add STLU
Date: Tue, 12 Jul 2011 17:13:24 +0800

commit 56db45a5cd06e3a6a7823a8cd7541e6bafe8427b upstream

For assembly code there are several "long" load and store defines already.
The one that's missing is the typical stack store, stdu/stwu.

So let's add that define as well, making my KVM code happy.

CC: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Alexander Graf <agraf@suse.de>
Acked-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Avi Kivity <avi@redhat.com>
Integrated-by: Tiejun Chen <tiejun.chen@windriver.com>
]

2/5 [
Author: Ananth N Mavinakayanahalli
Email: ananth@in.ibm.com
Subject: powerpc/kprobes: Remove resume_execution() in kprobes
Date: Tue, 12 Jul 2011 17:13:25 +0800

commit db97bc7f9997fef41b24c91f61c2f776d32d7ce3 upstream

emulate_step() in kprobe_handler() would've already determined if the
probed instruction can be emulated. We single-step in hardware only if
the instruction couldn't be emulated. resume_execution() therefore is
superfluous -- all we need is to fix up the instruction pointer after
single-stepping.

Thanks to Paul Mackerras for catching this.

Signed-off-by: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Integrated-by: Tiejun Chen <tiejun.chen@windriver.com>
]

3/5 [
Author: Paul Mackerras
Email: paulus@samba.org
Subject: powerpc: Emulate most Book I instructions in emulate_step()
Date: Tue, 12 Jul 2011 17:13:26 +0800

commit 0016a4cf5582415849fafbf9f019dd9530824789 upsream

This extends the emulate_step() function to handle a large proportion
of the Book I instructions implemented on current 64-bit server
processors.  The aim is to handle all the load and store instructions
used in the kernel, plus all of the instructions that appear between
l[wd]arx and st[wd]cx., so this handles the Altivec/VMX lvx and stvx
and the VSX lxv2dx and stxv2dx instructions (implemented in POWER7).

The new code can emulate user mode instructions, and checks the
effective address for a load or store if the saved state is for
user mode.  It doesn't handle little-endian mode at present.

For floating-point, Altivec/VMX and VSX instructions, it checks
that the saved MSR has the enable bit for the relevant facility
set, and if so, assumes that the FP/VMX/VSX registers contain
valid state, and does loads or stores directly to/from the
FP/VMX/VSX registers, using assembly helpers in ldstfp.S.

Instructions supported now include:
* Loads and stores, including some but not all VMX and VSX instructions,
  and lmw/stmw
* Atomic loads and stores (l[dw]arx, st[dw]cx.)
* Arithmetic instructions (add, subtract, multiply, divide, etc.)
* Compare instructions
* Rotate and mask instructions
* Shift instructions
* Logical instructions (and, or, xor, etc.)
* Condition register logical instructions
* mtcrf, cntlz[wd], exts[bhw]
* isync, sync, lwsync, ptesync, eieio
* Cache operations (dcbf, dcbst, dcbt, dcbtst)

The overflow-checking arithmetic instructions are not included, but
they appear not to be ever used in C code.

This uses decimal values for the minor opcodes in the switch statements
because that is what appears in the Power ISA specification, thus it is
easier to check that they are correct if they are in decimal.

If this is used to single-step an instruction where a data breakpoint
interrupt occurred, then there is the possibility that the instruction
is a lwarx or ldarx.  In that case we have to be careful not to lose the
reservation until we get to the matching st[wd]cx., or we'll never make
forward progress.  One alternative is to try to arrange that we can
return from interrupts and handle data breakpoint interrupts without
losing the reservation, which means not using any spinlocks, mutexes,
or atomic ops (including bitops).  That seems rather fragile.  The
other alternative is to emulate the larx/stcx and all the instructions
in between.  This is why this commit adds support for a wide range
of integer instructions.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Integrated-by: Tiejun Chen <tiejun.chen@windriver.com>
]

4/5 [
Author: Tiejun Chen
Email: tiejun.chen@windriver.com
Subject: booke/kprobe: make program exception to use one dedicated exception stack
Date: Thu, 14 Jul 2011 13:37:39 +0800

When kprobe targets operations such as store-and-update-word for SP(r1),

stwu r1, -A(r1)

A program exception is triggered and PPC always allocates an exception frame
as shown as the follows:

old r1 ----------
         ...
         nip
         gpr[2] ~ gpr[31]
         gpr[1] <--------- old r1 is stored.
         gpr[0]
       -------- <--------- pr_regs @offset 16 bytes
       padding
       STACK_FRAME_REGS_MARKER
       LR
       back chain
new r1 ----------
Then emulate_step() will emulate this instruction, 'stwu'. Which is
equivalent to:
1> Update pr_regs->gpr[1] = mem[old r1 + (-A)]
2> stw [old r1], mem[old r1 + (-A)]

Please notice the stack based on new r1 may be covered with mem[old r1
+(-A)] when addr[old r1 + (-A)] < addr[old r1 + sizeof(an exception frame0].
So operation #2 above operation will overwirte something to break this exception
frame then unexpected kernel behavior will be issued.

To fix this, we have to implement independent interrupt stack for PPC program
exception when CONFIG_BOOKE is enabled. Here we can use
EXC_LEVEL_EXCEPTION_PROLOG to replace original NORMAL_EXCEPTION_PROLOG
for program exception if CONFIG_BOOKE. Then its always safe for kprobe
with independent exc stack from one pre-allocated and dedicated thread_info.
This is similar to the implementation for critical/machine check/debug
exceptions on PPC.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

CQID: WIND00286005

5/5 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: oprofile: powerpc: deal signal frame as root one
Date: Tue, 12 Jul 2011 14:21:33 +0800

The stack frame of signal handler was constructed by kernel in user space and
its return address was set to vdso. Thus make the backtrace sample unresolved.
So check this condition and deal signal frame as root one to avoid those
unresolved symbols.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>